src/services/face_swap.py
from PIL import Image
from typing import Optional
from .config import config

# --- ВАЖНО ---
# Комментируем все импорты insightface, onnxruntime и OpenCV 
# чтобы не вызывать ошибку ImportError в окружении MVP.
# import cv2
# import insightface
# from insightface.app import FaceAnalysis

class FaceSwapper:
    """
    Класс для замены лица на изображении. 
    На этапе MVP - это заглушка.
    """
    def __init__(self):
        self.enabled = config.FACE_SWAP_ENABLED
        if self.enabled:
            # Тут будет логика инициализации InsightFace и ONNX моделей
            print("FaceSwapper: Активирована, требуется установка дополнительных библиотек!")
        else:
            print("FaceSwapper: Отключена в конфигурации.")
            
        # if self.enabled:
        #     self.app = FaceAnalysis(name='buffalo_l')
        #     self.app.prepare(ctx_id=0, det_size=(640, 640))
        #     self.swapper = insightface.model_zoo.get_model('inswapper_128.onnx')

    def swap_face(self, template_image_path: str, user_avatar_path: str) -> Optional[str]:
        """
        Меняет лицо на шаблоне, используя аватар пользователя.
        """
        if not self.enabled:
            # Если отключено, просто возвращаем оригинальный путь
            return template_image_path
        
        print("FaceSwapper: Выполняется сложная операция замены лица...")
        
        # Тут должна быть реальная логика InsightFace
        # try:
        #     img_target = cv2.imread(template_image_path)
        #     img_source = cv2.imread(user_avatar_path)
        #     ... (Логика из предыдущего ответа)
        #     cv2.imwrite(output_path, res)
        #     return output_path
        # except Exception as e:
        #     print(f"Ошибка при замене лица: {e}. Возвращаем оригинал.")
        #     return template_image_path
        
        # Возвращаем оригинал, пока логика закомментирована
        return template_image_path
