requirements.txt
aiogram>=3.1.1
openai>=1.2.3
requests>=2.31.0
Pillow>=10.1.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-decouple>=3.8
# insightface==0.7.3
# onnxruntime
